Issue 78: Add an interface on the GetUser query
