+++ src/Khatovar/Bundle/UserBundle/Controller/Admin/EditController.php
@@ -23,10 +23,10 @@ declare(strict_types=1);
 
 namespace Khatovar\Bundle\UserBundle\Controller\Admin;
 
-use Khatovar\Bundle\UserBundle\Entity\Exception\UserDoesNotExist;
 use Khatovar\Bundle\UserBundle\Form\Factory\UserFormFactory;
 use Khatovar\Bundle\UserBundle\Form\Type\UserType;
-use Khatovar\Bundle\UserBundle\Query\GetUser;
+use Khatovar\Bundle\UserBundle\Query\Exception\UserDoesNotExist;
+use Khatovar\Bundle\UserBundle\Query\GetUserFromDatabase;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Twig\Environment;
@@ -38,7 +38,7 @@ use Twig\Environment;
  */
 class EditController
 {
-    /** @var GetUser */
+    /** @var GetUserFromDatabase */
     private $getUser;
 
     /** @var UserFormFactory */
@@ -48,11 +48,11 @@ class EditController
     private $twig;
 
     /**
-     * @param GetUser         $getUser
-     * @param UserFormFactory $userFormFactory
-     * @param Environment     $twig
+     * @param GetUserFromDatabase $getUser
+     * @param UserFormFactory     $userFormFactory
+     * @param Environment         $twig
      */
-    public function __construct(GetUser $getUser, UserFormFactory $userFormFactory, Environment $twig)
+    public function __construct(GetUserFromDatabase $getUser, UserFormFactory $userFormFactory, Environment $twig)
     {
         $this->getUser = $getUser;
         $this->userFormFactory = $userFormFactory;
