+++ src/Khatovar/Bundle/UserBundle/Controller/Admin/UpdateController.php
@@ -23,11 +23,11 @@ declare(strict_types=1);
 
 namespace Khatovar\Bundle\UserBundle\Controller\Admin;
 
-use Khatovar\Bundle\UserBundle\Entity\Exception\UserDoesNotExist;
 use Khatovar\Bundle\UserBundle\Event\UserEvents;
 use Khatovar\Bundle\UserBundle\Form\Factory\UserFormFactory;
 use Khatovar\Bundle\UserBundle\Form\Type\UserType;
-use Khatovar\Bundle\UserBundle\Query\GetUser;
+use Khatovar\Bundle\UserBundle\Query\Exception\UserDoesNotExist;
+use Khatovar\Bundle\UserBundle\Query\GetUserFromDatabase;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\EventDispatcher\GenericEvent;
@@ -48,7 +48,7 @@ use Twig\Environment;
  */
 class UpdateController
 {
-    /** @var GetUser */
+    /** @var GetUserFromDatabase */
     private $getUser;
 
     /** @var UserFormFactory */
@@ -73,7 +73,7 @@ class UpdateController
     private $twig;
 
     /**
-     * @param GetUser                  $getUser
+     * @param GetUserFromDatabase      $getUser
      * @param UserFormFactory          $userFormFactory
      * @param EventDispatcherInterface $eventDispatcher
      * @param RegistryInterface        $doctrine
@@ -83,7 +83,7 @@ class UpdateController
      * @param Environment              $twig
      */
     public function __construct(
-        GetUser $getUser,
+        GetUserFromDatabase $getUser,
         UserFormFactory $userFormFactory,
         EventDispatcherInterface $eventDispatcher,
         RegistryInterface $doctrine,
